+++ changelog.md
@@ -1,3 +1,15 @@
+## V4.0.4 i18n Phase_2 FR EN
+
+Branche `i18n/V4.0.4_i18n_Phase_2`
+
+- Encapsulage des chaines a traduire suites les 5 regles
+- Passage de tous les messages console en anglais définitivement (print("message"))
+- Passage de la langue des scripts en anglais (run_bluenotebook.sh et .bat)
+- Traduction via Gemini CLI des fichiers FR fr.ts et EN en.ts
+- Traduction de l'aide en ligne --> online_help.html
+
+
+
 ## V4.0.3 i18n Windows Etape 1
 
 `develop` branch `i18n/V4.0.3_i18n_Windows`